# api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Shopmon API
  description: Shopware monitoring application API
  version: 1.0.0
servers:
  - url: /api
    description: API base path
...
security:
  - bearerAuth: []
paths:
...
  /organizations/{orgId}/sso-providers:
    get:
      operationId: getSsoProviders
      tags: [SSO]
      summary: List SSO providers for an organization
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: List of SSO providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SsoProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /sso/discover:
    get:
      operationId: discoverSso
      tags: [SSO]
      summary: Discover OIDC configuration from an issuer URL
      parameters:
        - name: issuer
          in: query
          required: true
          schema:
            type: string
            format: uri
      responses:
        "200":
          description: Discovered OIDC configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SsoDiscovery"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /organizations/{orgId}/sso-providers/{providerId}:
    put:
      operationId: updateSsoProvider
      tags: [SSO]
      summary: Update an SSO provider
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSsoProviderRequest"
      responses:
        "200":
          description: SSO provider updated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deleteSsoProvider
      tags: [SSO]
      summary: Delete an SSO provider
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "204":
          description: SSO provider deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  # Admin
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication

  # ──────────────────────────────────────────────
  # Parameters
  # ──────────────────────────────────────────────
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
      description: Organization ID
...
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
      description: SSO provider ID
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ──────────────────────────────────────────────
  # Schemas
  # ──────────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    # ── Account ──────────────────────────────────
...
    SsoProvider:
      type: object
      required:
        - id
        - domain
        - issuer
        - clientId
        - authorizationEndpoint
        - tokenEndpoint
        - jwksEndpoint
      properties:
        id:
          type: string
        domain:
          type: string
        issuer:
          type: string
        clientId:
          type: string
        authorizationEndpoint:
          type: string
        tokenEndpoint:
          type: string
        jwksEndpoint:
          type: string
    SsoDiscovery:
      type: object
      required:
        - issuer
        - authorizationEndpoint
        - tokenEndpoint
        - jwksEndpoint
        - userInfoEndpoint
        - scopes
      properties:
        issuer:
          type: string
        authorizationEndpoint:
          type: string
        tokenEndpoint:
          type: string
        jwksEndpoint:
          type: string
        userInfoEndpoint:
          type: string
        scopes:
          type: array
          items:
            type: string
    UpdateSsoProviderRequest:
      type: object
      required:
        - domain
        - issuer
        - clientId
        - authorizationEndpoint
        - tokenEndpoint
        - jwksEndpoint
      properties:
        domain:
          type: string
        issuer:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        authorizationEndpoint:
          type: string
        tokenEndpoint:
          type: string
        jwksEndpoint:
          type: string

    # ── Admin ────────────────────────────────────
...